+++ tests/covalent_tests/results_manager_tests/results_manager_test.py
@@ -25,18 +25,28 @@ from unittest import result
 from unittest.mock import ANY, Mock, call
 
 import pytest
+from sqlalchemy import create_engine
+from sqlalchemy.orm import Session
 
+from covalent._data_store import models
+from covalent._data_store.datastore import DataStore
 from covalent._data_store.models import Lattice
 from covalent._results_manager import wait
-from covalent._results_manager.results_manager import _get_result_from_dispatcher, result_from
+from covalent._results_manager.results_manager import (
+    _get_result_from_dispatcher,
+    result_from,
+    sync,
+)
 from covalent._shared_files.config import get_config
+
+DISPATCH_ID = "91c3ee18-5f2d-44ee-ac2a-39b79cf56646"
 
 
 @pytest.fixture
 def lattice_record():
     return Lattice(
         id=1,
-        dispatch_id="mock_dispatch_id",
+        dispatch_id=DISPATCH_ID,
         name="mock_name",
         status="mock_status",
         electron_num="mock_electron_num",
@@ -56,6 +66,17 @@ def lattice_record():
         started_at=None,
         completed_at=None,
     )
+
+
+class MockDataStore:
+    def __init__(self, lattice, dbpath):
+        engine = create_engine("sqlite+pysqlite:///" + str(dbpath / "workflow_db.sqlite"))
+        models.Base.metadata.create_all(engine)
+        session = Session(engine)
+        if lattice:
+            session.add(lattice)
+        session.commit()
+        self.engine = engine
 
 
 def test_result_from(lattice_record, mocker):
@@ -92,4 +113,31 @@ def test_get_result_from_dispatcher(mocker):
             ),
         ]
         * retries
+    )
+
+
+def test_sync(lattice_record, mocker, tmp_path):
+    mock_get_result_from_dispatcher = Mock()
+    mocker.patch(
+        "covalent._results_manager.results_manager._get_result_from_dispatcher",
+        mock_get_result_from_dispatcher,
+    )
+    db = MockDataStore(lattice_record, tmp_path)
+
+    dispatch_id = "9d1b308b-4763-4990-ae7f-6a6e36d35893"
+    sync(db, dispatch_id)
+    mock_get_result_from_dispatcher.assert_called_once_with(
+        dispatch_id, wait=wait.EXTREME, status_only=True
+    )
+    mock_get_result_from_dispatcher.reset_mock()
+    dispatch_ids = ["f34671d1-48f2-41ce-89d9-9a8cb5c60e5d", "02dfd929-d96b-4ad1-8411-a818ce465169"]
+    sync(db, dispatch_ids)
+    for dispatch_id in dispatch_ids:
+        mock_get_result_from_dispatcher.assert_any_call(
+            dispatch_id, wait=wait.EXTREME, status_only=True
+        )
+    mock_get_result_from_dispatcher.reset_mock()
+    sync(db)
+    mock_get_result_from_dispatcher.assert_called_once_with(
+        DISPATCH_ID, wait=wait.EXTREME, status_only=True
     )